fix: replace read_to_string with BufReader streaming in event replay (#657)
Replace `fs::read_to_string()` + in-memory parse with `BufReader::new(File::open()?)
.lines()` in `replay_events()` and `migrate_from_jsonl()`. Each JSONL line is
processed and dropped immediately — no full-file allocation at startup.

Additionally fix an infinite-loop bug on macOS: `File::open()` succeeds on
directories, causing `BufReader::lines()` to spin forever on EISDIR. Added a
metadata check after open to return Err for non-regular-file paths, and changed
the I/O error branch from `continue` to `break` to prevent unbounded loops on
unrecoverable stream errors.

Adds 13 new tests covering streaming correctness, CRLF handling, corruption
threshold logic, and the unreadable-path error path.
